ООП

+++ Examples/coursej/src/main/java/ru/mirea/oop/practice/coursej/api/vk/AudioApi.java
@@ -3,7 +3,7 @@ package ru.mirea.oop.practice.coursej.api.vk;
 import ru.mirea.oop.practice.coursej.api.vk.entities.Audio;
 
 public interface AudioApi extends ExternalCall {
-    Audio[] list(Long idOwner,
+    Audio[] list(Integer idOwner,
                  Integer idAlbum,
                  String idsAudio,
                  Integer needUser,
